% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTRfuncs.R
\name{add_dc_membership}
\alias{add_dc_membership}
\title{Add dynamic community membership to graphs as vertex attribute}
\usage{
add_dc_membership(allnets, dcmembership)
}
\arguments{
\item{allnets}{A list of igraph networks.}

\item{dcmembership}{A list of dynamic community membership as produced by
\code{\link{get_dc_membership}}}
}
\value{
The provided list of igraph networks with the added vertex attribute
   "DC" indicating which dynamic community that vertex is currently a member
   of.
}
\description{
Utility function for adding dynamic community memberships to a list of igraph
    networks
}
\examples{
data(allnets)
data(allcoms)
track = do_track(allnets, allcoms, history=1)
dcmembership = get_dc_membership(track)
#add vertex attribute to allnets, overwriting original allnets
allnets = add_dc_membership(allnets, dcmembership)
}
