% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTdata.R
\docType{data}
\name{allcoms}
\alias{allcoms}
\title{Generated example detected communities.}
\format{
A list of 6 numeric vectors, giving the community ID to which that
    vertex is assigned:
}
\usage{
allcoms
}
\description{
Detected communities from the dataset of network data to demonstrate
    MajorTrack (\code{\link{allnets}}). See vignette for code.
}
\keyword{datasets}
