% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTRfuncs.R
\name{community_lifespans}
\alias{community_lifespans}
\title{Get dynamic community lifespans}
\usage{
community_lifespans(track)
}
\arguments{
\item{track}{A MajorTrack object as produced by \code{\link{do_track}}.}
}
\value{
Named vector giving how many timesteps each dynamic community
    is present for.
}
\description{
Summary of how many timesteps each dynamic community in a MajorTrack object
    exists for.
}
\examples{
data(allnets)
data(allcoms)
track = do_track(allnets, allcoms, history=1)
community_lifespans(track)
}
