% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTRfuncs.R
\name{do_track}
\alias{do_track}
\title{Run MajorTrack algorithm}
\usage{
do_track(allnets, allcoms, historypar = 2)
}
\arguments{
\item{allnets}{A list of igraph networks.}

\item{allcoms}{A list of detected communities, as returned by many igraph
community detection algorithms. The list should be the same length as
allnets. Each element of the list should be a vector of integers with the
same length as the number of vertices in the corresponding network.}
}
\value{
A MajorTrack object, which R can interpret as a list. See
  \url{https://majortrack.readthedocs.io/en/latest/api/majortrack.tracker.html}
  for more details
}
\description{
\code{do_track} runs the MajorTrack algorithm for detecting dynamic communities
}
\examples{
data(allnets)
data(allcoms)
do_track(allnets, allcoms, history=1)
}
