% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTRfuncs.R
\name{get_similarities}
\alias{get_similarities}
\title{Get similarity in membership between communities}
\usage{
get_similarities(track)
}
\arguments{
\item{track}{A MajorTrack object as produced by \code{\link{do_track}}.}
}
\value{
Dataframe of 6 columns.

    The first column 'timestep' is the current timestep, to which
    timesteps forwards and backwards are being compared (except in the case
    of the first timestep).

    The second column 'group1' is the dynamic community to which
    dynamic communities in other timesteps are being compared.

    The third column 'direction' shows which direction the comparison is
    being made, either 'forwards' or 'backwards'

    The fourth column 'timestep2' indicates in which other timestep the
    comparison in being made.

    The fifth column 'group2' is the dynamic community in 'timestep2' to to
    which the dynamic community 'group1' is being compared to in 'timestep'

    The sixth column shows the similarity in membership between the two
    dynamic communities as a proportion. A value of 0 indicates that the
    compared communities contain none of the same members while a value of 1
    means the communities are exactly the same.
}
\description{
Summary of the similarity in membership between dynamic communities in
    different timesteps, both forwards and backwards.
}
\examples{
data(allnets)
data(allcoms)
track = do_track(allnets, allcoms, history=1)
get_similarities(track)
}
