% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTRfuncs.R
\name{ind_membership_df}
\alias{ind_membership_df}
\title{Get summaries of individual dynamic community membership}
\usage{
ind_membership_df(track = NULL, dcmembership = NULL)
}
\arguments{
\item{track}{A MajorTrack object as produced by \code{\link{do_track}}.
Not required if dcmembership is provided.}

\item{dcmembership}{A list of dynamic community membership as produced by
\code{\link{get_dc_membership}}. Not required if track is provided.}
}
\value{
A list consisting of two objects, each showing individual dynamic
  community membership in a different way.

   The first object 'memdf1' is a dataframe giving timestep, individual ID and
   dynamic community membership as 3 columns. This is useful for computation.

   The second object 'memdf2' is a matrix with a row for
   each individual and a column for each timestep.
   This is an easily readable way of looking at how an individual's membership
   changes over time.
}
\description{
Generate summaries of individual vertices' dynamic community membership over
    time.
}
\examples{
data(allnets)
data(allcoms)
track = do_track(allnets, allcoms, history=1)
indmembership=ind_membership_df(track)

#use a dummy variable to look at how many IDs are in each combination of timestep and groupsize

groupsizes=aggregate(rep(1,nrow(indmembership$memdf1))~
    group+timestep,FUN=sum,data=indmembership$memdf1,drop=F)
names(groupsizes)[3]="groupsize"
groupsizes$groupsize[is.na(groupsizes$groupsize)]=0
groupsizes

#matrix of individual memberships
head(indmembership$memdf2)
}
