% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MajorTrackR.R
\docType{package}
\name{MajorTrackR}
\alias{MajorTrackR}
\title{MajorTrackR: R interface for MajorTrack}
\description{
R interface for the majortrack python library for carrying out dynamic
    community detection method tracking persistent but transiently
    discontinuous communities in time window graphs.
}
\section{Functions}{

\code{\link{do_track}},
\code{\link{get_dc_membership}},
\code{\link{add_dc_membership}},
\code{\link{move_events_df}},
\code{\link{ind_membership_df}},
\code{\link{community_lifespans}},
\code{\link{get_similarities}},
\code{\link{get_alluvialplot}},
}

\section{Data}{

\code{\link{allnets}},
}

