% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTRfuncs.R
\name{get_dc_membership}
\alias{get_dc_membership}
\title{Get individual dynamic community membership from a MajorTrack object}
\usage{
get_dc_membership(track)
}
\arguments{
\item{track}{A MajorTrack object as produced by \code{\link{do_track}}}
}
\value{
A list, the length of which is the number of timesteps included in
   the MajorTrack object. Each element of the list is a numeric vector
   indicating which dynamic community a vertex is currently a member of.
}
\description{
\code{get_dc_membership} extracts the per timestep dynamic communities for
   each vertex.
}
\examples{
data(allnets)
#detect each networks communities
coms=lapply(allnets,function(x){
  igraph::cluster_louvain(x)
})
track = do_track(allnets, coms, history=1)
get_dc_membership(track)
}
